fix: 도메인 전반적으로 수정

+++ src/main/java/uos/uos25/Employee/DTO/EmployeeDTO.java
@@ -17,7 +17,7 @@ public class EmployeeDTO {
     @NotBlank(message = "주민번호를 입력하세요.")
     private String registrationNumber;
     @Min(value = -1, message = "0 이상의 값을 입력하세요.")
-    private Long salary;
+    private Integer salary;
     @NotBlank(message = "업무 시간을 입력하세요.")
     private String partTime;
     @NotBlank(message = "계좌번호를 입력하세요.")
